HiCoffeeDatabase/requirements.txt
asgiref==3.5.0
asttokens==2.0.5
autopep8 @ file:///opt/conda/conda-bld/autopep8_1639166893812/work
backcall==0.2.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.3
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
djoser==2.1.0
dnspython==2.2.1
drf-nested-routers==0.93.4
drf-yasg==1.20.0
executing==0.8.3
fpdf==1.7.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
ipython==8.1.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mongoengine==0.24.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pdfminer.six==20220319
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pure-eval==0.2.2
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1636635402688/work
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.11.2
PyJWT==2.3.0
pymongo==3.12.3
PyMySQL==1.0.2
pyparsing==3.0.8
python-docx==0.8.11
python-snappy==0.6.1
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.2.4
stack-data==0.2.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
whitenoise==6.0.0
wincertstore==0.2
winkerberos==0.9.0
